test: test for pipeline reentrance
